requirements.txt
pygame
pyyaml
python-box
tqdm